Move valid tests; drop old duplicated signer zero [skip ci]
