+++ types/spectest/tests/partialsigmessage/partial_root_valid.go
@@ -1,4 +1,4 @@
-package messages
+package partialsigmessage
 
 import (
 	"github.com/bloxapp/ssv-spec/qbft"
@@ -7,14 +7,14 @@ import (
 	"github.com/bloxapp/ssv-spec/types/testingutils"
 )
 
-// SigValid tests SignedPostConsensusMessage sig == 96 bytes
-func SigValid() tests.SpecTest {
+// PartialRootValid tests PostConsensusMessage root == 32 bytes
+func PartialRootValid() tests.SpecTest {
 	ks := testingutils.Testing4SharesSet()
 
 	msg := testingutils.PostConsensusAttestationMsg(ks.Shares[1], 1, qbft.FirstHeight)
 
 	return &MsgSpecTest{
-		Name: "sig valid",
+		Name: "partial root valid",
 		Messages: []*types.SignedPartialSignatureMessage{
 			msg,
 		},
