--- types/spectest/tests/partialsigmessage/signer_zero.go
@@ -1,6 +1,0 @@
-package partialsigmessage
-
-// SignerZero tests SignedPartialSignatureMessage signer ID 0
-func SignerZero() *MsgSpecTest {
-	panic("implement")
-}
